refactor: remove unscoped file from core index and update references, refs PM-2491
