+++ apps/web/src/app/core/index.ts
@@ -1,5 +1,4 @@
 export * from "./core.module";
 export * from "./event.service";
-export * from "../admin-console/core/policy-list.service";
 export * from "./router.service";
 export * from "./state/state.service";
